Add whatsNew attribute to OverviewCodec, add label attribute to NewsItemCodec and fix populate in overview fetch
