+++ .changeset/khaki-avocados-reflect.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": major
+---
+
+Add whatsNew attribute to OverviewCodec, add label attribute to NewsItemCodec and fix populate in overview fetch
