+++ apps/nextjs-website/src/lib/strapi/fetches/fetchOverviews.ts
@@ -39,6 +39,9 @@ const makeStrapiOverviewsPopulate = () =>
       bannerLinks: {
         populate: ['icon'],
       },
+      whatsNew: {
+        populate: ['link', 'items.image', 'items.link'],
+      },
     },
   });
